% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_ops.R
\name{axis.vector}
\alias{axis.vector}
\title{Unit vector for k-th axis in n-dimensional Euclidean space}
\usage{
axis.vector(n, k, column = !byrow, byrow = FALSE)
}
\arguments{
\item{n}{dimensionality of the Euclidean space}

\item{k}{which axis vector to return}

\item{column, byrow}{alternative flags to select a column vector (default) or row vector}
}
\value{
A \code{n x 1} matrix (default) or \code{1 x n} matrix (with \code{byrow=TRUE})
}
\description{
Returns a unit vector for the \code{k}-th axis in \code{n}-dimensional Euclidean space
in the form of a column vector (or row vector if specified by the user).
}
\examples{
axis.vector(5, 2)
axis.vector(5, 2, byrow=TRUE)
}
