% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cohen.d}
\alias{cohen.d}
\title{Compute Cohen's d as scale-adjusted effect size of a t-test}
\usage{
cohen.d(x, y, conf.level = NULL)
}
\arguments{
\item{x, y}{two numeric vectors giving the data points to be compared}

\item{conf.level}{if specified, return confidence interval of \eqn{d} at the specified
confidence level in range \eqn{[0, 1]} rather than point estimate}
}
\value{
Point estimate or confidence interval for Cohen's \eqn{d}.
}
\description{
This function computes Cohen's \eqn{d} as a scale-adjusted measure of effect size
for a t-test of two independent samples \code{x} and \code{y},
according to https://en.wikipedia.org/wiki/Effect_size#Cohen's_d
}
\details{
Effect size is based on the difference \eqn{\mu_x - \mu_y} in order to be compatible with
\code{\link[=t.test]{t.test()}}. Positive effect size thus indicates that \code{x} values are larger than \code{y} values on average.
}
