% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{expand.range}
\alias{expand.range}
\title{Expand range by a specified proportion on either side}
\usage{
expand.range(x, by = 0.2)
}
\arguments{
\item{x}{a numeric vector or matrix whose range is to be extended}

\item{by}{proportion by which to expand the range on either side (default: 20\%)}
}
\value{
A numeric vector of length 2 with the expanded range.
}
\description{
This function computes the range of the specified numerical values,
then expands it on either side by a specified proportion of the full range.
Its main use is to pick a suitable range for scatterplots, ensuring a margin around all data points.
}
\examples{
expand.range(1:10, by=.1)
}
