% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_ops.R
\name{gma.is.orthogonal}
\alias{gma.is.orthogonal}
\title{Check orthogonality of a column matrix}
\usage{
gma.is.orthogonal(M, fail = FALSE, tol = 1e-07)
}
\arguments{
\item{M}{matrix of one or more numeric column vectors}

\item{fail}{if TRUE, abort with error message if the matrix is not orthogonal}

\item{tol}{tolerance of orthogonality test (default: \eqn{10^{-7}})}
}
\value{
\code{TRUE} or \code{FALSE} (unless the function aborts because of \code{fail=TRUE}).
}
\description{
This function checks whether a given column matrix is orthogonal, i.e. whether
its column vectors are orthonormal.
}
