% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{gma.lda}
\alias{gma.lda}
\title{Repeated-measures Linear Discriminant Analysis}
\usage{
gma.lda(
  x,
  grouping,
  cohorts = NULL,
  tol = 1e-04,
  balanced = FALSE,
  R2 = 100,
  ...
)
}
\arguments{
\item{x}{row matrix of data vectors (as in \code{\link[MASS:lda]{MASS::lda()}})}

\item{grouping}{factor sepcifying group membership of each data point (as in \code{\link[MASS:lda]{MASS::lda()}})}

\item{cohorts}{optional factor specifying cohort membership for each data point.
If specified, a repeated-measures LDA is carried out that factors cohorts out of within-group variance.}

\item{tol}{minimum condition number required for covariance matrix.
Unlike \code{\link[MASS:lda]{MASS::lda()}}, we do not yet support the case of a singular covariance matrix.}

\item{balanced}{if TRUE, all groups make the same contribution to the between-group covariance matrix (regardless of group size)}

\item{R2}{choose as many discriminants as needed to cover at least this percentage of relative group separation variance (default: 100\%)}

\item{...}{any further argument are ignored, but allowed for compatibiltiy with \code{\link[MASS:lda]{MASS::lda()}}}
}
\value{
An object of class \code{lda} which mimics the value returned by \code{\link[MASS:lda]{MASS::lda()}}. In particular, the object includes components
\itemize{
\item \code{scaling}: column matrix of LDA axis vectors; multiply with data matrix to obtain LDA dimension scores
\item \code{prior}: the prior probabilities of class membership used by the algorithm (depending on the \code{balanced} flag)
\item \code{means}: the group means in the original space (which can be transformed into LDA scores with \code{scaling})
\item \code{svd}: singular values for the ratio of between- and within-group variance, from which \code{R2} values are derived
\item \code{counts}: number of data points in each group
\item \code{lev}: group labels (the levels of \code{grouping})
}
}
\description{
Our own implementation of linear discriminant analysis (LDA), including
a novel repeated-measures version in order to remove confounding factors
from within-group variance. For a complete description of the algorithm
and its mathematical details, see
\href{https://github.com/schtepf/GMA/blob/main/doc/gma_maths.pdf}{Mathematics of GMA}.
}
