% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_ops.R
\name{gma.orthogonalize}
\alias{gma.orthogonalize}
\title{Gram-Schmidt orthonormalisation of a list of vectors}
\usage{
gma.orthogonalize(M, fail = TRUE)
}
\arguments{
\item{M}{matrix of one or more numeric column vectors}

\item{fail}{if \code{fail=FALSE} is specified, the function will silently return \code{NULL}
if the vectors don't have full rank (rather than aborting with an error).}
}
\value{
A matrix of the same shape as \code{M} with the orthonormalised vectors (or \code{NULL}).
}
\description{
This function applies Gram-Schmidt orthonormalisation to a list of vectors
(given as columns of matrix \code{M}) via QR decomposition. This algorithm ensures
that the span of the first \eqn{k} vectors remains the same after orthonormalisation.
}
\examples{
tmp <- matrix(1, 5, 3)
tmp <- tmp * (row(tmp) <= col(tmp))
tmp
gma.orthogonalize(tmp)
}
