% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_plot.R
\name{gma.pairs}
\alias{gma.pairs}
\title{Scatterplot Matrix for GMA Focus Spaces}
\usage{
gma.pairs(
  M,
  dims = NULL,
  Meta = NULL,
  select = NULL,
  alpha.select = 0,
  pch = NULL,
  col = NULL,
  pch.vals = 1:10,
  pch.cols = NULL,
  col.vals = corpora.palette("simple"),
  cex = 1,
  legend.cex = 1.4 * cex,
  randomize = TRUE,
  gap = 0.5,
  oma = c(2, 2, 2, 2),
  iso = FALSE,
  compact = FALSE,
  lim = NULL,
  ...
)
}
\arguments{
\item{M}{a row matrix of data points to be plotted}

\item{dims}{which dimensions of \code{M} to include in the scatterplot matrix (default: all)}

\item{Meta}{a data frame of metadata corresponding to the data points in \code{Meta}}

\item{select}{expression to select a subset of data points to be plotted, either as logical vector or vector of row numbers.
The expression is evaluated within \code{Meta}, so its columns can directly be used as variables. Note that \code{select} only
reduces data points to a subset but does not reorder them.}

\item{alpha.select}{if set to a positive value in the range \eqn{(0, 1]}, data points omitted by a \code{select} expression are not
hidden completely, but their opacity is reduced to the specified value}

\item{pch}{expression to determine plot symbols for the data points, usually based on columns of \code{Meta} (which can be accessed directly as variables)}

\item{col}{expression to determine colours for the data points, usually based on columns of \code{Meta} (which can be accessed directly as variables)}

\item{pch.vals}{plot symbols used for the levels of the expression \code{pch} (values are looked up by level index, not by level name)}

\item{pch.cols}{option colours for the levels of the expression \code{pch}, so plot symbols can appear in different colours (cannot be combined with \code{col})}

\item{col.vals}{point colours used for the levels of the expression \code{col} (values are looked up by level index, not by level name)}

\item{cex}{global magnification factor for plot symbols and text}

\item{legend.cex}{magnification factor for legend boxes (not relative to \code{cex}, but default is \code{1.4 * cex})}

\item{randomize}{if TRUE (the default), data points are plotted in random order so no categories are hidden.
Specify a positive integer to set a random seed for reproducibility, or set to FALSE to keep original plotting order.}

\item{gap}{distance between panels of the scatter plot matrix, in margin lines (default: 0.5)}

\item{oma}{adjusted outer margins of individual panels for a more compact display (internal use)}

\item{iso}{if TRUE, adjust ranges for all axes so that the scatterplots are isometric (provided that the overall plot has square aspect).
That is, ranges have the same width on each axis, but are shifted to center the data.}

\item{compact}{if TRUE, only the upper triangle of scatterplots is shown and legends are moved to the lower triangle, for a more compact display}

\item{dim}{use \code{lim=c(min, max)} to specify a fixed range for all coordinate axes,
or choose individual ranges as an \eqn{n_{\text{dim}} \times 2}{n_dim x 2} matrix (overrides \code{iso=TRUE})}
}
\description{
\strong{TODO:}
write a description
}
