% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weights.R
\name{gma.plot.weights}
\alias{gma.plot.weights}
\title{Barplot of feature weights for one or more focus space dimensions}
\usage{
gma.plot.weights(
  basis,
  dim = seq_len(ncol(basis)),
  idx = NULL,
  names = NULL,
  feature.names = NULL,
  ylab = "normalized feature weights",
  main = NULL,
  bw = FALSE,
  base_size = 12,
  ylim = NULL,
  zlim = NULL
)
}
\arguments{
\item{basis}{column matrix of basis vectors for the focus space dimensions}

\item{dim}{dimensions to be shown in plot (default: all). If more than one dimension is selected, the barplots will be stacked vertically.}

\item{idx}{optional index vector selecting a subset of feature weights to be displayed}

\item{names, feature.names}{= labels for dimensions and features in the plot. Labels must be provided for selected dimensions only,
but for all features (before selection with \code{idx}).}

\item{ylab}{label for y-axis of the plot}

\item{main}{optional main title for the plot}

\item{bw}{if TRUE, produce a black & white version of the plot}

\item{base_size}{basic font size for the plot (see \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw()}} documentation)}

\item{ylim}{display range on y-axis (NULL or two-element vector)}

\item{zlim}{range of values mapped to the colour scale, with out-of-band values clamped to maximal colours.
It is always ensured that the colour range is symmetric around 0 so that zero weights are white.}
}
\value{
A \code{ggplot} object that has to be printed to display the graph.
}
\description{
\strong{TODO: description}
The \strong{ggplot2} package must be installed in order to use this function.
}
