% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMA.R
\name{GMA}
\alias{GMA}
\title{R6 Class for Geometric Multivariate Analysis}
\value{
Unless otherwise specified, methods invisibly return the object so they can be chained.
}
\description{
\strong{TODO}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{P}}{orthonormal basis of focus space (column matrix)}

\item{\code{Q}}{orthonormal basis of complement space (column matrix)}

\item{\code{axes}}{basis spanning focus space, as specified by user (column matrix)}

\item{\code{data}}{data matrix of feature vectors in original coordinates (row matrix)}

\item{\code{n}}{dimensionality of the feature space}

\item{\code{rotated}}{TRUE if rotation has been applied, which invalidates \code{axes}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GMA-new}{\code{GMA$new()}}
\item \href{#method-GMA-print}{\code{GMA$print()}}
\item \href{#method-GMA-basis}{\code{GMA$basis()}}
\item \href{#method-GMA-projection}{\code{GMA$projection()}}
\item \href{#method-GMA-R2}{\code{GMA$R2()}}
\item \href{#method-GMA-add}{\code{GMA$add()}}
\item \href{#method-GMA-subspace.apply}{\code{GMA$subspace.apply()}}
\item \href{#method-GMA-discriminant}{\code{GMA$discriminant()}}
\item \href{#method-GMA-add.discriminant}{\code{GMA$add.discriminant()}}
\item \href{#method-GMA-similarity}{\code{GMA$similarity()}}
\item \href{#method-GMA-rotation}{\code{GMA$rotation()}}
\item \href{#method-GMA-clone}{\code{GMA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-new"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-new}{}}}
\subsection{Method \code{new()}}{
Create a new GMA object. \strong{TODO}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$new(M = NULL, n.dim = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{numeric matrix of feature vectors (as rows) representing the main data set to be analysed}

\item{\code{n.dim}}{dimensionality of the feature space, to be specified if \code{M} is not provided}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{GMA} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-print"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-print}{}}}
\subsection{Method \code{print()}}{
Print a brief description of a GMA object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-basis"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-basis}{}}}
\subsection{Method \code{basis()}}{
Retrieve basis of focus space and/or complement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$basis(space = c("focus", "both", "complement", "axes"), dim = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{space}}{select desired basis vectors to be returned
\itemize{
\item \code{focus}: orthonormal basis of focus space (default)
\item \code{complement} orthonormal basis of complement space
\item \code{both}: orthonormal basis of focus space + complement
\item \code{axes}: user-specified basis of focus space (need not be orthonormal)
}}

\item{\code{dim}}{= return only these dimensions of the selected basis}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A column matrix of basis vectors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-projection"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-projection}{}}}
\subsection{Method \code{projection()}}{
Project data points into subspace and/or complement
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$projection(
  space = c("focus", "both", "complement", "axes"),
  dim = NULL,
  M = NULL,
  byrow = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{space}}{select subspace and type of projection
\itemize{
\item \code{focus}: orthogonal projection into focus space (default)
\item \code{complement} orthogonal projection into complement space
\item \code{both}: transform into orthonormal basis of focus space + complement
\item \code{axes}: focus space coordinates in user-specified basis
}}

\item{\code{dim}}{= return only these coordinates of the selected projection}

\item{\code{M}}{= data matrix of row vectors to project (defaults to internal data matrix)}

\item{\code{byrow}}{set \code{byrow=FALSE} if \code{M} is a matrix of column vectors}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix of row vectors with the coordinates of data points after projection (or column vectors for \code{byrow=FALSE}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-R2"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-R2}{}}}
\subsection{Method \code{R2()}}{
Compute proportion of variance (\eqn{R^2}) captured by projection into focus space.
The return value is a vector of \eqn{R^2} contributions (\%) for each selected dimension.
Note that additional complement dimensions can be included with the \code{dim} argument.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$R2(M = NULL, dim = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{row matrix of data points to be projected (defaults to original data matrix)}

\item{\code{dim}}{= vector of dimensions for which \eqn{R^2} is computed (defaults to complete focus space)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector specifying partial \eqn{R^2} for each selected dimensions as a percentage.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-add"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-add}{}}}
\subsection{Method \code{add()}}{
Add one or more basis dimensions to the focus space. The user-specified basis vectors must be
a linearly independent extension of the existing focus space. They are automatically transformed
into an orthonormal basis extension of the focus space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$add(
  basis,
  space = c("orig", "both", "complement"),
  names = colnames(basis),
  normalize = FALSE,
  tol = 1e-07
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{basis}}{column matrix of basis vectors to be added (a single basis vector can be given as plain vector)}

\item{\code{space}}{orthonormal coordinate system in which basis vectors are specified
\itemize{
\item \code{orig}: original coordinates (feature space, the default)
\item \code{both}: coordinates in orthonormal basis (of focus space + complement)
\item \code{complement}: coordinates within complement space
}}

\item{\code{names}}{optional names for the additional dimensions}

\item{\code{normalize}}{normalise specified basis vectors to unit length if \code{normalize=TRUE}}

\item{\code{tol}}{tolerance for internal orthogonality checks}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-subspace.apply"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-subspace.apply}{}}}
\subsection{Method \code{subspace.apply()}}{
Apply operation in subspace coordinates and re-transform resulting vectors into original feature space.
Only orthogonal transformations are allowed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$subspace.apply(
  M,
  FUN,
  ...,
  space = c("focus", "both", "complement"),
  dim = NULL,
  byrow.in = byrow,
  byrow.out = byrow,
  byrow = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{data matrix of row or column vectors to be operated on}

\item{\code{FUN}}{function applied to projected data matrix}

\item{\code{...}}{any additional arguments are passed to \code{FUN} after the transformed data matrix}

\item{\code{space}}{select desired subspace (space, complement, both), defaults to "space"
\itemize{
\item \code{focus}: orthogonal projection into focus space (default)
\item \code{complement} orthogonal projection into complement space
\item \code{both}: transform into orthonormal basis of focus space + complement
}}

\item{\code{dim}}{use only these dimensions of the selected subspace}

\item{\code{byrow.in}}{whether input \code{M} is a row or column matrix}

\item{\code{byrow.out}}{whether result of \code{FUN} is a row or column matrix}

\item{\code{byrow}}{default value for both \code{byrow.in} and \code{byrow.out} (defaults to TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A row (if \code{byrow.out=TRUE}) or column matrix of vectors returned by \code{FUN}, transformed back into the original feature space.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-discriminant"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-discriminant}{}}}
\subsection{Method \code{discriminant()}}{
Identify LDA dimensions in full space or specified subspace.
Note that the dimensions returned are LDA discriminants rather than orthonormal basis vectors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$discriminant(
  categories,
  cohorts = NULL,
  balanced = FALSE,
  R2 = 100,
  M = NULL,
  space = c("both", "complement", "focus"),
  dim = NULL,
  idx = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{vector of categories to be separated by LDA (must match rows of \code{M})}

\item{\code{cohorts}}{optional vector of cohort assignments for repeated-measures LDA (see \code{\link[=gma.lda]{?gma.lda}})}

\item{\code{balanced}}{set \code{balanced=TRUE} to give all groups the same contribution to between-group variance (regardless of group size)}

\item{\code{R2}}{what percentage of relative between-group variance needs to be captured by discriminant space}

\item{\code{M}}{data matrix of row vectors (defaults to internal data matrix of GMA object)}

\item{\code{space}}{select desired subspace in which LDA is carried out
\itemize{
\item \code{both}: transform into orthonormal basis of focus space + complement (default)
\item \code{complement} orthogonal projection into complement space
\item \code{focus}: orthogonal projection into focus space
}}

\item{\code{dim}}{use only these dimensions of the selected subspace}

\item{\code{idx}}{perform LDA on subset of M selected by \code{idx}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
column matrix of LDA discriminants (in coordinates of original feature space)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-add.discriminant"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-add.discriminant}{}}}
\subsection{Method \code{add.discriminant()}}{
Extend focus space with LDA dimensions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$add.discriminant(
  categories,
  cohorts = NULL,
  balanced = FALSE,
  R2 = 100,
  M = NULL,
  space = c("complement", "both"),
  idx = NULL,
  max.dim = Inf
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{vector of categories to be separated by LDA (must match rows of \code{M})}

\item{\code{cohorts}}{optional vector of cohort assignments for repeated-measures LDA (see \code{\link[=gma.lda]{?gma.lda}})}

\item{\code{balanced}}{set \code{balanced=TRUE} to give all groups the same contribution to between-group variance (regardless of group size)}

\item{\code{R2}}{what percentage of relative between-group variance needs to be captured by discriminant space}

\item{\code{M}}{data matrix of row vectors (defaults to internal data matrix of GMA object)}

\item{\code{space}}{select desired subspace in which LDA is carried out
\itemize{
\item \code{complement} orthogonal projection into complement space (default)
\item \code{both}: transform into orthonormal basis of focus space + complement
}}

\item{\code{idx}}{perform LDA on subset of M selected by \code{idx}}

\item{\code{max.dim}}{maximum number of LDA dimensions to add to the focus space}

\item{\code{dim}}{use only these dimensions of the selected subspace}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-similarity"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-similarity}{}}}
\subsection{Method \code{similarity()}}{
Compute similarity between the focus spaces of two GMA objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$similarity(
  other,
  method = c("dimensions", "R2", "sigma"),
  add.basis = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{a second GMA object based on the same underlying feature space}

\item{\code{method}}{similarity measure to compute (\code{dim}, \code{R2}, or \code{sigma}); see \link[=gma.similarity]{?gma.similarity} for details}

\item{\code{add.basis}}{if TRUE, attach matching basis vectors as attributes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The selected similarity score, or a vector of cosine similarities between matching dimensions (for \code{method="sigma"}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-rotation"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-rotation}{}}}
\subsection{Method \code{rotation()}}{
Apply a rotation to (selected dimensions of) the focus space.
In contrast to factor analysis, all GMA rotations are orthogonal transformations that preserve the geometry of the focus space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$rotation(
  type = c("pca", "swap", "flip", "auto", "turn", "manual", "align"),
  dim = NULL,
  M = NULL,
  perm = NULL,
  basis = NULL,
  phi = 0,
  tol = 1e-06,
  check = TRUE,
  debug = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{selects the rotation to be performed:
\itemize{
\item \code{pca}:    perform a PCA on \code{M}(or internal data matrix) in the subspace and rotate the dimensions to principal components (default)
\item \code{swap}:   reorder dimensions according to the permutation in \code{perm} (default: reverse order)
\item \code{flip}:   reverse sign of selected dimensions
\item \code{auto}:   swap and flip dimensions to give best match to those in \code{basis}
\item \code{turn}:   rotate two basis dimensions of focus space by user-specified angle \code{phi} (counter-clockwise, note that projected data points rotate clockwise)
\item \code{manual}: rotate basis vectors to match manually specified axes (in \code{basis}), up to differences in sign
\itemize{
\item \code{basis} can specify fewer axes than the dimensionality of the subspace, use permutation in \code{dim} to select which basis vectors are matched
\item axes are projected into focus space (or subspace selected by \code{dim}) and (re-)orthogonalised
}
\item \code{align}:  align with subspace in \code{basis} by finding optimally matching orthonormal basis vectors in both spaces
\itemize{
\item uses \code{\link[=gma.similarity]{gma.similarity()}} algorithm; both subspaces must have the same dimensionality
\item apply twice to align two GMA focus spaces: \code{A$rotation("align", basis=B)} and \code{B$rotation("align", basis=A)}
}
}}

\item{\code{dim}}{subset of the focus space dimensions on which rotation will be performed
(since the dimensions are orthogonal, rotations on disjoint subsets are completely independent)}

\item{\code{M}}{optional data matrix for \code{pca} rotation (defaults to internal data)}

\item{\code{perm}}{a permutation of the selected focus space dimensions (\code{1:d}), defaults to \code{rev(1:d)}}

\item{\code{basis}}{a column matrix of basis vectors (for \code{auto}, \code{manual}, \code{align}) or a GMA object (whose focus space basis will be used)}

\item{\code{phi}}{manually specified angle for \code{turn} rotation (in degrees)}

\item{\code{tol, check, debug}}{internal use arguments for diagnostics and debugging}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
CAVEAT: rotations invalidate the user-specified basis vectors (\verb{$axes}), which are replaced with the orthonormal basis vectors \verb{$P}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GMA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
