% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace_similarity.R
\name{gma.similarity}
\alias{gma.similarity}
\title{Compute similarity between two subspaces A and B}
\usage{
gma.similarity(
  A,
  B,
  method = c("dimensions", "R2", "sigma"),
  orthogonalize = TRUE,
  tol = 1e-07,
  basis = FALSE
)
}
\arguments{
\item{A}{column matrix of basis vectors of subspace A}

\item{B}{column matrix of basis vectors of subspace B}

\item{method}{similarity measure: number of shared dimensions (\code{dim}), preserved variance (\code{R2}),
or vector of singular values for matched basis vectors (\code{sigma})}

\item{orthogonalize}{can be set to FALSE if \code{A} and \code{B} are already orthnormal bases}

\item{tol}{tolerance for testing orthogonality in this case (default: \eqn{10^{-7}})}

\item{basis}{matched basis vectors for the two subspaces are attached as attributes if \code{basis=TRUE}}
}
\value{
Either a numeric similarity score or a vector of singular values. If \code{basis=TRUE},
attributes \code{basisA} and \code{basisB} provide column matrices of matched basis vectors
}
\description{
Computes various measures of the similarity of two subspaces \code{A} and \code{B},
which are meaningful even if the dimensionalities of the two subspaces differ.
}
\details{
Subspace similarity is based on an SVD of the crossproduct of the two sets of basis vectors,
which determines orthogonal transformations of both bases such that corresponding basis
vectors are optimally matched. The singular values are cosine similarities between pairs of
matched basis vectors. Depending on \code{method}, subspace similarity is quantified as:
\itemize{
\item \code{dim}: sum of the singular values, which can be understood as the fractional number of
shared dimensions between the subspaces. Its range is \eqn{[0, \min \{\dim A, \dim B\}]}.
\item \code{R2}: average proportion of preserved variance if random vectors from B are projected into A.
Its range is \eqn{[0, 1]}.
\item \code{sigma}: vector of individual singular values for matched basis vectors
If both subspaces have the same dimensionality, then all measures are symmetric. Otherwise,
similarity is computed for a projection from B into A.
}

Further information and some mathematical details are sketched in \href{https://sigil.r-forge.r-project.org/#unit07}{SIGIL Unit 7}.
}
