\name{gmatools-package}
\alias{gmatools-package}
\alias{gmatools}
\docType{package}
\title{\packageTitle{gmatools}}
\description{\packageDescription{gmatools}}
\section{Package Content}{\packageIndices{gmatools}}
\author{\packageAuthor{gmatools}}
\section{Maintainer}{\packageMaintainer{gmatools}}
\keyword{package}
