% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{signed.log}
\alias{signed.log}
\title{Signed logarithmic transformation that is smooth at 0}
\usage{
signed.log(x, base = exp(1))
}
\arguments{
\item{x}{a numeric vector or matrix containing the values to be transformed}

\item{base}{base \eqn{b} of the logarithm to apply (defaults to \eqn{e})}
}
\value{
A numeric vector or matrix of the same shape with the logarithmic transformation applied.
}
\description{
This function applies a signed logarithmic transformation to the argument
using the mapping \deqn{x \mapsto \mathop{\text{sgn}}(x) \cdot \log_b(1 + |x|)}{x -> sgn(x) * log_b(1 + |x|)}
}
\examples{
signed.log(seq(-3, 3), base=2)
\dontrun{
plot(signed.log, from=-4, to=4, xlim=c(-4, 4), ylim=c(-4, 4),
     lwd=3, col=2, xaxs="i", yaxs="i")
abline(h=0, v=0, lwd=2); abline(0, 1, col=4)}
}
