\name{Gmatools-package}
\alias{Gmatools-package}
\alias{Gmatools}
\docType{package}
\title{
\packageTitle{Gmatools}
}
\description{
\packageDescription{Gmatools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Gmatools}
\packageIndices{Gmatools}
This package contains some functions to merge output from MCMCglmm package and estimate classic metrics used to compare Gmatrix
}
\author{
\packageAuthor{Gmatools}

Maintainer: \packageMaintainer{Gmatools}
}
\references{

}

\keyword{ package }
\seealso{

\code{\link[<MCMCglmm>:<MCMCglmm>-package]{<MCMCglmmg>}} 
}
\examples{

}
